refactor(ui): fix import paths after directory restructuring, extract paste/history hooks
- Fix stale relative import paths in cli.tsx, ConfigDropdown, DropdownMenu,
  McpStatusList, UpdatePrompt, and App.tsx after src/ui/ directory restructuring
- Extract handlePaste/expandPasteMarkerAtCursor into usePasteHandling hook
- Extract navigateHistory into useHistoryNavigation hook
- Clean up unused imports in App.tsx, move resolveCurrentSettings locally
