+++ src/ui/views/PromptInput.tsx
@@ -604,106 +604,6 @@ export const PromptInput = React.memo(function PromptInput({
     });
   }
 
-  function handlePaste(pastedText: string): void {
-    const totalChars = pastedText.length;
-
-    if (totalChars <= 1000) {
-      const newlineCount = (pastedText.match(/\n/g) ?? []).length;
-      if (newlineCount <= 9) {
-        const clean = pastedText
-          .replace(/\r\n|\r/g, "\n")
-          .replace(/[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]/g, "")
-          .replace(/\t/g, "    ");
-        updateBuffer((s) => insertText(s, clean));
-        return;
-      }
-    }
-
-    // Large paste: store raw text, insert marker with line/char count.
-    const lineCount = (pastedText.match(/\n/g) ?? []).length + 1;
-    pasteCounterRef.current += 1;
-    const pasteId = pasteCounterRef.current;
-    pastesRef.current.set(pasteId, pastedText);
-
-    const marker =
-      lineCount > 10 ? `[paste #${pasteId} +${lineCount} lines]` : `[paste #${pasteId} ${totalChars} chars]`;
-
-    updateBuffer((s) => insertText(s, marker));
-  }
-
-  function expandPasteMarkerAtCursor(): void {
-    // First, try to collapse an already-expanded region at the cursor.
-    for (const [id, region] of expandedRegionsRef.current) {
-      if (buffer.cursor >= region.start && buffer.cursor <= region.end) {
-        // Collapse back to marker.
-        expandedRegionsRef.current.delete(id);
-        pastesRef.current.set(id, region.content);
-        setTimeout(() => {
-          updateBuffer((s) => {
-            const text = s.text.slice(0, region.start) + region.marker + s.text.slice(region.end);
-            return { text, cursor: region.start + region.marker.length };
-          });
-        }, 0);
-        return;
-      }
-    }
-
-    // No expanded region at cursor — try to expand a paste marker.
-    const marker = findPasteMarkerContaining(buffer);
-    if (!marker) {
-      setStatusMessage(t("ui.promptInput.noPasteMarker"));
-      return;
-    }
-    const content = pastesRef.current.get(marker.id);
-    if (!content) {
-      setStatusMessage(t("ui.promptInput.pasteNotFound"));
-      return;
-    }
-
-    const pasteId = marker.id;
-    const originalMarker = buffer.text.slice(marker.start, marker.end);
-    pastesRef.current.delete(pasteId);
-
-    setTimeout(() => {
-      updateBuffer((s) => {
-        const text = s.text.slice(0, marker.start) + cleanPasteContent(content) + s.text.slice(marker.end);
-        const newEnd = marker.start + content.length;
-        expandedRegionsRef.current.set(pasteId, {
-          start: marker.start,
-          end: newEnd,
-          content,
-          marker: originalMarker,
-        });
-        return { text, cursor: marker.start };
-      });
-    }, 0);
-  }
-
-  function navigateHistory(direction: -1 | 1): void {
-    if (promptHistory.length === 0) {
-      return;
-    }
-
-    const previousCursor = historyCursor === -1 ? promptHistory.length : historyCursor;
-    const nextCursor = Math.max(0, Math.min(promptHistory.length, previousCursor + direction));
-    const draft = historyCursor === -1 ? buffer.text : draftBeforeHistory;
-
-    if (historyCursor === -1) {
-      setDraftBeforeHistory(buffer.text);
-    }
-
-    if (nextCursor === promptHistory.length) {
-      const text = draft ?? "";
-      setBuffer({ text, cursor: text.length });
-      setHistoryCursor(-1);
-      setDraftBeforeHistory(null);
-      return;
-    }
-
-    const text = promptHistory[nextCursor] ?? "";
-    setBuffer({ text, cursor: text.length });
-    setHistoryCursor(nextCursor);
-  }
   function insertFileMentionSelection(item: FileMentionItem): void {
     if (!fileMentionToken) {
       return;
